x86: support for tracing J_DOUBLE and J_FLOAT return value.

+++ include/jit/compiler.h
@@ -120,6 +120,6 @@ void trace_exception_handler(struct compilation_unit *, unsigned char *);
 void trace_exception_unwind(struct jit_stack_frame *);
 void trace_exception_unwind_to_native(struct jit_stack_frame *);
 void trace_bytecode(struct vm_method *);
-void trace_return_value(struct vm_method *, unsigned long);
+void trace_return_value(struct vm_method *, unsigned long long);
 
 #endif
